# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DJ-IT API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/saved-views": {
      "get": {
        "tags": [
          "saved-views"
        ],
        "summary": "List Saved Views",
        "operationId": "list_saved_views_api_v1_saved_views_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedViewSummary"
                  },
                  "type": "array",
                  "title": "Response List Saved Views Api V1 Saved Views Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "saved-views"
        ],
        "summary": "Create Saved View",
        "operationId": "create_saved_view_api_v1_saved_views_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/saved-views/{view_id}": {
      "patch": {
        "tags": [
          "saved-views"
        ],
        "summary": "Update Saved View",
        "operationId": "update_saved_view_api_v1_saved_views__view_id__patch",
        "parameters": [
          {
            "name": "view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "View Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "saved-views"
        ],
        "summary": "Delete Saved View",
        "operationId": "delete_saved_view_api_v1_saved_views__view_id__delete",
        "parameters": [
          {
            "name": "view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "View Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Delete Saved View Api V1 Saved Views  View Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/saved-views/{view_id}/freeze": {
      "post": {
        "tags": [
          "saved-views"
        ],
        "summary": "Freeze Saved View",
        "operationId": "freeze_saved_view_api_v1_saved_views__view_id__freeze_post",
        "parameters": [
          {
            "name": "view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "View Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewFreezeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewFreezeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/export/engine-dj": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Export Engine Dj",
        "operationId": "export_engine_dj_api_v1_export_engine_dj_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "type": "object",
                  "title": "Response Export Engine Dj Api V1 Export Engine Dj Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/copy-to-folder": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Export Copy To Folder",
        "description": "Copy all playlist track files into <destination_path>/<playlist_name>/.",
        "operationId": "export_copy_to_folder_api_v1_export_copy_to_folder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCopyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportCopyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/{playlist_id}/extract": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Extract Playlist",
        "operationId": "extract_playlist_api_v1_playlists__playlist_id__extract_post",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCopyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportCopyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportCopyRequest": {
        "properties": {
          "playlist_id": {
            "type": "integer",
            "title": "Playlist Id"
          },
          "destination_path": {
            "type": "string",
            "title": "Destination Path"
          }
        },
        "type": "object",
        "required": [
          "playlist_id",
          "destination_path"
        ],
        "title": "ExportCopyRequest"
      },
      "ExportCopyResult": {
        "properties": {
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "copied": {
            "type": "integer",
            "title": "Copied"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "provisional": {
            "type": "integer",
            "title": "Provisional",
            "default": 0
          },
          "playlist_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playlist Path"
          },
          "manifest_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Path"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "destination",
          "copied",
          "skipped",
          "errors"
        ],
        "title": "ExportCopyResult"
      },
...
      "ExportRequest": {
        "properties": {
          "playlist_id": {
            "type": "integer",
            "title": "Playlist Id"
          },
          "engine_library_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Library Path"
          }
        },
        "type": "object",
        "required": [
          "playlist_id"
        ],
        "title": "ExportRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SavedViewCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "state": {
            "$ref": "#/components/schemas/SavedViewState"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "state"
        ],
        "title": "SavedViewCreate"
      },
      "SavedViewFreezeRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "SavedViewFreezeRequest"
      },
      "SavedViewFreezeResponse": {
        "properties": {
          "playlist_id": {
            "type": "integer",
            "title": "Playlist Id"
          },
          "playlist_name": {
            "type": "string",
            "title": "Playlist Name"
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count"
          }
        },
        "type": "object",
        "required": [
          "playlist_id",
          "playlist_name",
          "track_count"
        ],
        "title": "SavedViewFreezeResponse"
      },
      "SavedViewState": {
        "properties": {
          "search": {
            "type": "string",
            "title": "Search",
            "default": ""
          },
          "folder_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Path"
          },
          "triage_decision": {
            "items": {
              "type": "string",
              "enum": [
                "unheard",
                "keep",
                "maybe",
                "reject",
                "problem"
              ]
            },
            "type": "array",
            "title": "Triage Decision"
          },
          "analysis_status": {
            "items": {
              "type": "string",
              "enum": [
                "not_analyzed",
                "pending",
                "analyzing",
                "done",
                "failed",
                "skipped",
                "overridden"
              ]
            },
            "type": "array",
            "title": "Analysis Status"
          },
          "key_camelot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Camelot"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "energy_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Min"
          },
          "energy_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Max"
          },
          "bpm_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm Min"
          },
          "bpm_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm Max"
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "default": "triage_decision"
          },
          "sort_desc": {
            "type": "boolean",
            "title": "Sort Desc",
            "default": false
          }
        },
        "type": "object",
        "title": "SavedViewState"
      },
      "SavedViewSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "$ref": "#/components/schemas/SavedViewState"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "SavedViewSummary"
      },
      "SavedViewUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavedViewState"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "type": "object",
        "title": "SavedViewUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}